SE/CompareIt/requirements.txt
black==23.11.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
flake8==6.1.0
Flask==3.0.0
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
psycopg==3.1.13
psycopg-binary==3.1.13
pycodestyle==2.11.1
pyflakes==3.1.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
python-dotenv==1.0.0